Moved the view for the ServerSide of the Xelon table in the Squalaetp app
