+++ api/tests/tests_api_views.py
@@ -58,16 +58,3 @@ class ApiTestCase(APITestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.data), 4)
         self.assertEqual(response.data, {"count": 0, "next": None, "previous": None, "results": []})
-
-    def test_xelon_view_set_is_disconnected(self):
-        response = self.client.get(reverse('api:xelon-list'), format='json')
-        self.assertEqual(response.status_code, 403)
-        self.assertEqual(response.data, self.authError)
-
-    # def test_xelon_view_set_is_connected(self):
-    #     self.login('admin')
-    #     response = self.client.get(reverse('api:xelon-list'), format='json')
-    #     self.assertEqual(response.status_code, 200)
-    #     self.assertEqual(len(response.data), 4)
-
-
