+++ api/urls.py
@@ -8,7 +8,6 @@ router.register(r'users', views.UserViewSet)
 router.register(r'groups', views.GroupViewSet)
 router.register(r'prog', views.ProgViewSet, basename='prog')
 router.register(r'cal', views.CalViewSet, basename='cal')
-router.register(r'xelon', views.XelonViewSet, basename='xelon')
 router.register(r'unlock', views.UnlockViewSet, basename='unlock')
 router.register(r'reman/batch', views.RemanBatchViewSet, basename='reman_batch')
 router.register(r'reman/checkout', views.RemanCheckOutViewSet, basename='reman_checkout')
