+++ psa/templates/psa/corvet_table.html
@@ -4,7 +4,7 @@
 
 {% block table_content %}
 
-    <table class="table table-bordered table-sm text-nowrap" id="corvetNewTable" width="100%" cellspacing="0">
+    <table class="table table-bordered table-sm text-nowrap display" id="corvetNewTable" style="width: 100%;">
         <thead>
         <tr>
             <th></th>
