+++ squalaetp/views.py
@@ -11,7 +11,11 @@ from django.views.generic import TemplateView
 from bootstrap_modal_forms.generic import BSModalUpdateView, BSModalFormView
 from django.forms.models import model_to_dict
 from django.core.management import call_command
+from rest_framework.response import Response
+from rest_framework import viewsets, permissions, status
 
+from api.models import QueryTableByArgs
+from .serializers import XelonSerializer, XELON_COLUMN_LIST
 from .models import Xelon, Stock, Action
 from .utils import collapse_select
 from psa.models import Corvet
@@ -229,3 +233,34 @@ def change_table(request):
     table_title = 'Historique des changements Squalaetp'
     actions = Action.objects.all()
     return render(request, 'squalaetp/change_table.html', locals())
+
+
+class XelonViewSet(viewsets.ModelViewSet):
+    permission_classes = (permissions.IsAuthenticated,)
+    queryset = Xelon.objects.filter(date_retour__isnull=False)
+    serializer_class = XelonSerializer
+
+    def list(self, request, **kwargs):
+        try:
+            self._filter(request)
+            xelon = QueryTableByArgs(self.queryset, XELON_COLUMN_LIST, 2, **request.query_params).values()
+            serializer = self.serializer_class(xelon["items"], many=True)
+            data = {
+                "data": serializer.data,
+                "draw": xelon["draw"],
+                "recordsTotal": xelon["total"],
+                "recordsFiltered": xelon["count"],
+            }
+            return Response(data, status=status.HTTP_200_OK)
+        except Exception as err:
+            return Response(err, status=status.HTTP_404_NOT_FOUND)
+
+    def _filter(self, request):
+        query = request.query_params.get('filter', None)
+        if query and query == 'pending':
+            self.queryset = self.queryset.exclude(type_de_cloture__in=['Réparé', 'N/A'])
+        elif query and query == "vin-error":
+            self.queryset = self.queryset.filter(vin_error=True).order_by('-date_retour')
+        elif query and query == "corvet-error":
+            self.queryset = self.queryset.filter(
+                vin__regex=r'^VF[37]\w{14}$', vin_error=False, corvet__isnull=True).order_by('-date_retour')
